requirements.txt
mkdocs==1.6.1
mkdocs-get-deps==0.2.2
mkdocs-material==9.7.6
mkdocs-material-extensions==1.3.1
mkdocs-wikilinks-plugin==0.1.2
python-frontmatter==1.1.0
